Move HttpError back into http; and make tokio required
I noticed there's actually just a single function in gateway that can throw an HTTP error. So instead of moving the entire HTTP infrastructure into root, I just gate that one gateway function behind http instead

Also, I made tokio required because... you're gonna use it one way or another, and it was a bit annoying with the feature flags
