+++ src/internal/mod.rs
@@ -3,5 +3,4 @@ pub mod macros;
 
 pub mod prelude;
 
-#[cfg(feature = "tokio")]
 pub mod tokio;
